fix: address code review findings on Playwright E2E setup
- Extract TEST_USER_EMAIL/PASSWORD/NAME and DEFAULT_STAGES into
  e2e/constants.ts as single source of truth; import from both
  e2e/helpers.ts and backend/scripts/seed-test.ts to eliminate
  credential duplication that would cause silent login failures
- Add environment safety guard in seed-test.ts: refuse to TRUNCATE
  unless TEST_DATABASE_URL contains 'test', 'neon.tech', or 'localhost'
  to prevent accidental production wipe from a misconfigured env var
- Replace inline --compiler-options JSON in seed:test script with
  backend/tsconfig.seed.json for portability (Windows compatibility)
- Fix CI cache-dependency-path to reference only the root package-lock.json
  (backend/package-lock.json and frontend/package-lock.json are not committed)
- CI migrate step now calls npm run migrate instead of duplicating the
  ts-node invocation directly
- Capture backend startup logs to /tmp/backend.log in CI; dump log on
  wait-on failure so silent backend crashes are visible
- Add --host 0.0.0.0 to frontend preview for self-hosted runner compatibility
- Add permissions: contents: read to e2e workflow (least privilege)
- Replace fragile page.locator('h3').first() with a named Stage assertion
  (DEFAULT_STAGES[0]) in auth.spec.ts first test
- Add expect(page).not.toHaveURL('/login') before waitForURL in login()
  so failed logins produce a clear error instead of a 60s timeout
- Randomize JWT_SECRET fallback in test-e2e.sh with openssl rand -hex 32
  instead of a predictable hardcoded string
- Add explicit timeout: 60_000 to playwright.config.ts
